Remove ProjectManifest to file
